refactor producer collector
