+++ src/main/java/no/sysco/middleware/prometheus/kafka/clients/ProducerJmxCollector.java
@@ -2,15 +2,20 @@ package no.sysco.middleware.prometheus.kafka.clients;
 
 
 import io.prometheus.client.Collector;
+import io.prometheus.client.GaugeMetricFamily;
 import no.sysco.middleware.prometheus.kafka.common.KafkaClientJmxCollector;
 import no.sysco.middleware.prometheus.kafka.internal.ProducerMetricsTemplates;
-import org.apache.kafka.common.Metric;
 import org.apache.kafka.common.MetricName;
 import org.apache.kafka.streams.KeyValue;
 
-import javax.management.*;
+import javax.management.MBeanServer;
+import javax.management.MalformedObjectNameException;
+import javax.management.ObjectName;
 import java.lang.management.ManagementFactory;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -44,7 +49,7 @@ public class ProducerJmxCollector extends KafkaClientJmxCollector {
     // kafka.producer:type=producer-topic-metrics,client-id=25862c0e-9da0-48f7-82b5-cdf077d1ff6a,topic=topic-2
     public List<KeyValue<String, String>> getClientTopicList(final String clientId) {
         String objectNameWithDomain =
-                ProducerMetricsTemplates.PRODUCER_DOMAIN  +
+                ProducerMetricsTemplates.PRODUCER_DOMAIN +
                         ":type=" + ProducerMetricsTemplates.PRODUCER_TOPIC_METRIC_GROUP_NAME +
                         ",client-id=" + clientId + ",*";
         List<KeyValue<String, String>> clientTopicList = new ArrayList<>();
@@ -66,18 +71,44 @@ public class ProducerJmxCollector extends KafkaClientJmxCollector {
     public List<Collector.MetricFamilySamples> getMetrics() {
         List<Collector.MetricFamilySamples> metricsDefinedAtStart =
                 getMetricsPerClientId(ProducerMetricsTemplates.PRODUCER_METRIC_GROUP_NAME, producerMetricNamesAtStartup);
+        List<Collector.MetricFamilySamples> perTopicMetric = getPerTopicMetric();
+        return Stream
+                .concat(metricsDefinedAtStart.stream(), perTopicMetric.stream())
+                .collect(Collectors.toList());
+    }
 
+    public List<Collector.MetricFamilySamples> getPerTopicMetric() {
         Set<String> clientIds = getKafkaClientIds(ProducerMetricsTemplates.PRODUCER_METRIC_GROUP_NAME);
         List<KeyValue<String, String>> clientsTopicsList = new ArrayList<>();
         for (String id : clientIds) {
             List<KeyValue<String, String>> topicsPerClient = getClientTopicList(id);
             clientsTopicsList.addAll(topicsPerClient);
         }
-        Set<MetricName> metricsPerClientIdTopic  = producerMetricsTemplates.getMetricNamesClientIdTopic(clientsTopicsList);
+        Set<MetricName> metricsPerClientIdTopic = producerMetricsTemplates.getMetricNamesClientIdTopic(clientsTopicsList);
         List<Collector.MetricFamilySamples> metricsDefinedAtRuntime = getMetricsPerClientIdTopic(ProducerMetricsTemplates.PRODUCER_TOPIC_METRIC_GROUP_NAME, metricsPerClientIdTopic);
-        List<Collector.MetricFamilySamples> allMetrics = Stream.concat(metricsDefinedAtStart.stream(), metricsDefinedAtRuntime.stream())
-                .collect(Collectors.toList());
-        return allMetrics;
+        return metricsDefinedAtRuntime;
+    }
+
+    public List<Collector.MetricFamilySamples> getMetricsPerClientIdTopic(final String metricType, final Set<MetricName> metricNames) {
+        List<Collector.MetricFamilySamples> metricFamilySamples = new ArrayList<>();
+        for (MetricName metricName : metricNames) {
+            String clientId = metricName.tags().get("client-id");
+            String topic = metricName.tags().get("topic");
+
+
+            GaugeMetricFamily gaugeMetricFamily = new GaugeMetricFamily(
+                    formatMetricName(metricName),
+                    metricName.description(),
+                    Arrays.asList("client-id", "topic")
+            );
+            gaugeMetricFamily.addMetric(
+                    Arrays.asList(clientId, topic),
+                    getMBeanAttributeValue(metricType, metricName.name(), KeyValue.pair("client-id", clientId), KeyValue.pair("topic", topic))
+            );
+            System.out.println("HEREZ: "+gaugeMetricFamily);
+            metricFamilySamples.add(gaugeMetricFamily);
+        }
+        return metricFamilySamples;
     }
 
 }
